Refactor the `check_int` function to a better home.
