+++ portal/views/assessment_engine.py
@@ -42,7 +42,7 @@ from ..models.questionnaire_bank import QuestionnaireBank
 from ..models.role import ROLE
 from ..models.user import User, current_user, get_user_or_abort
 from ..trace import dump_trace, establish_trace
-from .portal import check_int
+from ..type_tools import check_int
 
 assessment_engine_api = Blueprint('assessment_engine_api', __name__,
                                   url_prefix='/api')
